+++ packages/orq-rc/src/models/operations/index.ts
@@ -171,6 +171,7 @@ export * from "./postv2feedbackevaluationremove.js";
 export * from "./postv2feedbackremove.js";
 export * from "./postv2humanevalsets.js";
 export * from "./postv2routerocr.js";
+export * from "./previewdatasourcechunks.js";
 export * from "./projectdelete.js";
 export * from "./projectget.js";
 export * from "./projectlist.js";