+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -182,7 +182,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-08-31T11:24:22.202Z",
+    "2026-09-01T06:04:38.639Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {